+++ src/gui/web-components/form/AircraftWebComponent.ts
@@ -1,6 +1,7 @@
 import type { AeroflyAircraft } from "@fboes/aerofly-data/data/aircraft-liveries.json";
 import { sendToMain } from "../../renderer/sendToMain.js";
 import { AbstractStateSubscriberWebComponent } from "../util/AbstractStateSubscriberWebComponent.js";
+import { registerElement } from "../util/registerElement.js";
 
 export type AircraftWebComponentState = {
     aircraftName: string;
@@ -24,8 +25,8 @@ export class AircraftWebComponent extends AbstractStateSubscriberWebComponent {
         return {
             aircraftName: this.elements.aircraftName.value,
             aircraftPaintscheme: this.elements.aircraftPaintscheme.value,
-            cruiseSpeed_kts: this.elements.aircraftCruiseSpeed.valueAsNumber,
-            cruiseAltitude_ft: this.elements.aircraftCruiseAltitude.valueAsNumber,
+            cruiseSpeed_kts: this.elements.aircraftCruiseSpeed.valueAsNumber || 1,
+            cruiseAltitude_ft: this.elements.aircraftCruiseAltitude.valueAsNumber || 1,
         };
     }
 
@@ -34,7 +35,7 @@ export class AircraftWebComponent extends AbstractStateSubscriberWebComponent {
         this.innerHTML = `\
 <div class="d-flex">
     <div class="form-group">
-        <label for="aircraft-name" class="header">✈️ Aircraft</label>
+        <label for="aircraft-name" class="header"><startgeraet-icon icon="airplane"></startgeraet-icon>&nbsp;Aircraft</label>
         <select id="aircraft-name">
             <option>Cessna 172</option>
         </select>
@@ -125,6 +126,6 @@ export class AircraftWebComponent extends AbstractStateSubscriberWebComponent {
     };
 
     static registerElement() {
-        customElements.define("startgeraet-aircraft", AircraftWebComponent);
+        registerElement("startgeraet-aircraft", AircraftWebComponent);
     }
 }