+++ src/gui/web-components/form/FlightPlanChooserWebComponent.ts
@@ -1,6 +1,7 @@
 import { dispatchNotificationEvent, type NotificationEventPayload } from "../../renderer/notificationEventHandler.js";
 import { sendToMain } from "../../renderer/sendToMain.js";
 import type { ImportWebComponentPayload } from "./ImportWebComponent.js";
+import { registerElement } from "../util/registerElement.js";
 
 export type FlightPlanChooserWebComponentState = {
     flightPlanIndex: number;
@@ -96,6 +97,6 @@ export class FlightPlanChooserWebComponent extends HTMLElement {
     };
 
     static registerElement() {
-        customElements.define("aerofly-flightplan-chooser", FlightPlanChooserWebComponent);
+        registerElement("aerofly-flightplan-chooser", FlightPlanChooserWebComponent);
     }
 }