+++ src/gui/web-components/form/FlightplanWebComponent.ts
@@ -1,4 +1,5 @@
 import { AbstractStateSubscriberWebComponent } from "../util/AbstractStateSubscriberWebComponent.js";
+import { registerElement } from "../util/registerElement.js";
 
 export class FlightplanWebComponent extends AbstractStateSubscriberWebComponent {
     private isInitialized = false;
@@ -13,7 +14,7 @@ export class FlightplanWebComponent extends AbstractStateSubscriberWebComponent 
     private initialize() {
         this.setAttribute("aria-role", "region");
         this.innerHTML = `\
-<h3>🛫 Flight plan</h3>
+<h3><startgeraet-icon icon="clipboard-check"></startgeraet-icon>&nbsp;Flight plan</h3>
 <table>
   <thead>
     <tr>
@@ -72,6 +73,6 @@ export class FlightplanWebComponent extends AbstractStateSubscriberWebComponent 
     }
 
     static registerElement() {
-        customElements.define("startgeraet-flightplan", FlightplanWebComponent);
+        registerElement("startgeraet-flightplan", FlightplanWebComponent);
     }
 }