+++ src/gui/web-components/form/ImportWebComponent.ts
@@ -1,6 +1,7 @@
 import { dispatchNotificationEvent, type NotificationEventPayload } from "../../renderer/notificationEventHandler.js";
 import { sendToMain } from "../../renderer/sendToMain.js";
 import { FlightPlanChooserWebComponent } from "./FlightPlanChooserWebComponent.js";
+import { registerElement } from "../util/registerElement.js";
 
 export type ImportWebComponentPayload = {
     flightplans: string[];
@@ -71,6 +72,6 @@ export class ImportWebComponent extends HTMLElement {
     };
 
     static registerElement() {
-        customElements.define("startgeraet-import", ImportWebComponent);
+        registerElement("startgeraet-import", ImportWebComponent);
     }
 }