+++ src/gui/web-components/form/TimeAndDateWebComponent.ts
@@ -1,5 +1,6 @@
 import { sendToMain } from "../../renderer/sendToMain.js";
 import { AbstractStateSubscriberWebComponent } from "../util/AbstractStateSubscriberWebComponent.js";
+import { registerElement } from "../util/registerElement.js";
 
 export type TimeAndDateWebComponentState = {
     utcDate: string; // YYYY-MM-DD
@@ -21,7 +22,7 @@ export class TimeAndDateWebComponent extends AbstractStateSubscriberWebComponent
     private initialize() {
         this.setAttribute("aria-role", "region");
         this.innerHTML = `\
-<h3>⏰ Time &amp; date</h3>
+<h3><startgeraet-icon icon="clock"></startgeraet-icon>&nbsp;Date &amp; time</h3>
 
 <table>
   <thead>
@@ -141,6 +142,6 @@ export class TimeAndDateWebComponent extends AbstractStateSubscriberWebComponent
     }
 
     static registerElement() {
-        customElements.define("startgeraet-time-and-date", TimeAndDateWebComponent);
+        registerElement("startgeraet-time-and-date", TimeAndDateWebComponent);
     }
 }