+++ src/gui/web-components/form/WeatherStatusWebComponent.ts
@@ -1,6 +1,7 @@
 import type { AeroflylightCategoryUs, AeroflylightCategoryIcao } from "../../../core/util/AeroflyFlightHelper.js";
 import { sendToMain } from "../../renderer/sendToMain.js";
 import { AbstractStateSubscriberWebComponent } from "../util/AbstractStateSubscriberWebComponent.js";
+import { registerElement } from "../util/registerElement.js";
 
 export class WeatherStatusWebComponent extends AbstractStateSubscriberWebComponent {
     private isInitialized = false;
@@ -15,7 +16,7 @@ export class WeatherStatusWebComponent extends AbstractStateSubscriberWebCompone
         this.innerHTML = `\
 <div class="d-flex">
     <div class="form-group">
-        <label for="flight-category-icao" class="header">⛅ Flight category</label>
+        <label for="flight-category-icao" class="header"><startgeraet-icon icon="cloud-sun"></startgeraet-icon>&nbsp;Flight category</label>
         <select id="flight-category-icao">
             <option value="VFR">VFR</option>
             <option value="IFR">IFR</option>
@@ -68,6 +69,6 @@ export class WeatherStatusWebComponent extends AbstractStateSubscriberWebCompone
     };
 
     static registerElement() {
-        customElements.define("startgeraet-weather-status", WeatherStatusWebComponent);
+        registerElement("startgeraet-weather-status", WeatherStatusWebComponent);
     }
 }