+++ src/gui/web-components/structure/FooterWebComponent.ts
@@ -1,3 +1,5 @@
+import { registerElement } from "../util/registerElement.js";
+
 export class FooterWebComponent extends HTMLElement {
     private isInitialized = false;
 
@@ -16,6 +18,6 @@ This application cannot set fuel, payload, runway waypoints, and the starting po
     }
 
     static registerElement() {
-        customElements.define("startgeraet-footer", FooterWebComponent);
+        registerElement("startgeraet-footer", FooterWebComponent);
     }
 }