+++ types/gui/web-components/util/registerElement.d.ts
@@ -1,0 +1,5 @@
+/**
+ * Register custom web component, preventing duplicate definitions.
+ */
+export declare function registerElement(name: string, constructor: CustomElementConstructor, options?: ElementDefinitionOptions): void;
+//# sourceMappingURL=registerElement.d.ts.map