+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Improved GUI for wide screen
 - Added more compact GUI
+- Added abstract icon web component
 
 ## [2.2.2] -2026-07-31